backend/finetune_llm/requirements.txt
# QLoRA Fine-Tuning Dependencies
# Install with: pip install -r requirements.txt

# Core ML frameworks
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0

# Quantization
bitsandbytes>=0.41.0

# Dataset handling
datasets>=2.15.0

# Training utilities
trl>=0.7.0
wandb>=0.16.0

# Model conversion
sentencepiece>=0.1.99
protobuf>=3.20.0

# GGUF conversion (for Ollama)
llama-cpp-python>=0.2.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
numpy>=1.24.0
scipy>=1.11.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.8.0

# HTTP client (for Ollama API)
httpx>=0.25.0
aiofiles>=23.0.0
